requirements.txt
# Core functionality
basil_daq>=2.4.3 # readout framework pyBAR is built on
pySiLibUSB # support for SiLab USB devices
pyBAR_fei4_interpreter>=1.2.0  # to interpret the raw data
pixel_clusterizer>=2.5  # to cluster pixel hits
bitarray>=0.8.1
matplotlib
numpy
progressbar-latest>=2.4
tables
pyyaml
scipy

# Additional functionality (lab devices, unit tests, Online Monitor)
pyvisa # interface to lab devices
pyvisa-py # interface to lab devices
pyserial # serial port
pyqtgraph # monitor
mock # unit tests
sphinx # for vitables
vitables # data viewer
nose # unit tests
pyzmq # monitor